[core] Fix: COM Archive FastObjectTypes had a problem and would break after a restart.
Former-commit-id: 540c8bee34f63137c834ac8b58709fd9a831191d

+++ core/MO_SERVICES_LAYER/CCSDS_COM_IMPL/src/main/java/esa/mo/com/impl/archive/db/COMObjectEntityPK.java
@@ -28,11 +28,11 @@ import java.io.Serializable;
  */
 public class COMObjectEntityPK implements Serializable {
 
-    private final Long objectTypeId;
+    private final Integer objectTypeId;
     private final Integer domainId;
     private final Long objId;
 
-    public COMObjectEntityPK (final Long objectTypeId, final Integer domain, final Long objId){
+    public COMObjectEntityPK (final Integer objectTypeId, final Integer domain, final Long objId){
         this.objectTypeId = objectTypeId;
         this.domainId = domain;
         this.objId = objId;
